+++ src/Input/Touchable.h
@@ -17,7 +17,7 @@ public:
 	/// User puts finger down on the device.
 	/// \param touches  The id and position of each touch
 	/// \param count    Number of touches
-	virtual void touch_began(const Touch *const touches, const unsigned int count) = 0;
+	virtual void touch_began(const Touch *const touches, const size_t count) = 0;
 
 	/// Canceled touch event. May occur when the user moves finger off-screen
 	/// or the application got interrupted by external events.
@@ -26,12 +26,12 @@ public:
 	/// User lifts finger.
 	/// \param touches  The id and position of each touch
 	/// \param count    Number of touches
-	virtual void touch_ended(const Touch *const touches, const unsigned int count) = 0;
+	virtual void touch_ended(const Touch *const touches, const size_t count) = 0;
 
 	/// User moves finger around while it still is on the device.
 	/// \param touches  The id and position of each touch
 	/// \param count    Number of touches
-	virtual void touch_moved(const Touch *const touches, const unsigned int count) = 0;
+	virtual void touch_moved(const Touch *const touches, const size_t count) = 0;
 };
 
 #endif
