+++ src/Lua/lua_Input.h
@@ -19,10 +19,10 @@ public:
 	static void key_up(lua_State *L, const Key &key);
 #endif
 
-	static void touch_began(lua_State *L, const Touch *const touches, const unsigned int count);
+	static void touch_began(lua_State *L, const Touch *const touches, const size_t count);
 	static void touch_canceled(lua_State *L);
-	static void touch_ended(lua_State *L, const Touch *const touches, const unsigned int count);
-	static void touch_moved(lua_State *L, const Touch *const touches, const unsigned int count);
+	static void touch_ended(lua_State *L, const Touch *const touches, const size_t count);
+	static void touch_moved(lua_State *L, const Touch *const touches, const size_t count);
 
 private:
 	static void getfield(lua_State *L, const char *const field);
@@ -32,7 +32,7 @@ private:
 	static void key_event(lua_State *L, const char *const type, const Key &key);
 #endif
 
-	static void touch_event(lua_State *L, const char *const type, const Touch *const touches = 0, const unsigned int count = 0);
+	static void touch_event(lua_State *L, const char *const type, const Touch *const touches = 0, const size_t count = 0);
 
 	lua_Input();
 	lua_Input(const lua_Input &);
